domains: use symmetric representation for FF of prime order
Before 754a316ea8, symmetric=True option was a default for
FF and some methods in the polys module (in galoistools and
so on).  Apparently, some remnants of that were left, e.g.
dup_trunc(), dmp_zz_modular_resultant(), PolyElement.trunc_ground()
and so on (see aaf07e78d4).  This commit restores symmetric=True
as a default (and only) view of FF elements for outside world
(e.g., this will be used in conversion to ints).